scripts/cloud_function/requirements.txt
functions-framework==3.*
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.0.0
PyPDF2>=3.0.0
vertexai>=1.38.0
cloudevents>=1.9.0